Organizing and Refactoring code in modules
