+++ .github/workflows/python-package.yml
@@ -40,4 +40,4 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Tests with unittest
       run: |
-        python test/tests.py
+        python -m unittest
