+++ brFinance/scraper/__init__.py
@@ -1,0 +1,6 @@
+import ssl
+import warnings
+
+
+ssl._create_default_https_context = ssl._create_unverified_context
+warnings.simplefilter(action='ignore', category=FutureWarning)
