+++ brFinance/utils/browser.py
@@ -1,20 +1,16 @@
-from selenium import webdriver
-import platform
-from fake_useragent import UserAgent
 import os
-from webdriver_manager.chrome import ChromeDriverManager
-
-from datetime import date
-import math
-from dateutil.relativedelta import relativedelta
-import time
-
-import pandas as pd
+import platform
+from time import sleep
 from dataclasses import dataclass
+from selenium import webdriver
+from webdriver_manager.chrome import ChromeDriverManager
+from fake_useragent import UserAgent
+
 
 @dataclass
 class Browser:
 
+    @staticmethod
     def download_wait(path_to_downloads):
         """
         Waits all Chrome download files (.crdownload) in a folder be done before continues
@@ -22,7 +18,7 @@ class Browser:
         seconds = 0
         dl_wait = True
         while dl_wait and seconds < 20:
-            time.sleep(1)
+            sleep(1)
             dl_wait = False
             for fname in os.listdir(path_to_downloads):
                 if fname.endswith('.crdownload'):
@@ -30,6 +26,7 @@ class Browser:
             seconds += 1
         return seconds
 
+    @staticmethod
     def run_chromedriver(hidden: bool=True) -> webdriver:
         """
         Instantiate a webdriver object with different settings depending of OS you are using
@@ -69,35 +66,3 @@ class Browser:
             driver = webdriver.Chrome(options=options)
 
         return driver
-
-
-@dataclass
-class File:
-    def __init__(self, directory: str) -> None:
-        self.directory = directory
-        
-    def check_exist(self) -> bool:
-        if os.path.isfile(self):
-            return True
-        else:
-            return False
-    
-    def create_folder(self) -> None:
-        try:
-            os.makedirs(self)
-        except FileExistsError:
-            # directory already exists
-            pass
-
-@dataclass
-class Dates:
-    def __init__(self, _date: date):
-        self.date = _date
-    
-    @property
-    def previous_quarter_end_date(self) -> date:
-        return (self.date - relativedelta(months=3) - pd.tseries.offsets.DateOffset(day=1) + pd.tseries.offsets.QuarterEnd()).date()
-
-
-if __name__ == '__main__':
-    print(Dates(date(2021, 7, 12)).previous_quarter_end_date)
